test and move some files
