+++ package.json
@@ -14,10 +14,12 @@
     "start": "tsdx watch",
     "build": "tsdx build",
     "test": "tsdx test --passWithNoTests",
+    "posttest": "npm run format",
     "lint": "tsdx lint",
     "prepare": "tsdx build",
     "size": "size-limit",
-    "analyze": "size-limit --why"
+    "analyze": "size-limit --why",
+    "format": "prettier --loglevel warn --write \"**/*.{ts,tsx,css,md}\""
   },
   "peerDependencies": {
     "react": ">=16"
@@ -50,18 +52,18 @@
     "@size-limit/preset-small-lib": "^7.0.5",
     "@types/react": "^17.0.38",
     "@types/react-dom": "^17.0.11",
+    "@types/styled-components": "^5.1.21",
     "husky": "^7.0.4",
+    "prettier": ">=2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "size-limit": "^7.0.5",
     "tsdx": "^0.14.1",
     "tslib": "^2.3.1",
-    "typescript": "^4.5.5",
-    "@types/styled-components": "^5.1.21"
+    "typescript": "^4.5.5"
   },
   "dependencies": {
     "@reach/dialog": "^0.16.2",
-
     "styled-components": "^5.3.3"
   }
 }
