+++ src/StyledModal/styles.ts
