+++ src/constants.ts
@@ -1,0 +1,3 @@
+
+export const PARAMS_KEY = 'params';
+export const MODAL_KEY = 'modal';
