--- test/blah.test.tsx
@@ -1,5 +1,0 @@
-describe('it', () => {
-  it('renders without crashing', () => {
-    expect(true).toBe(true);
-  });
-});
